+++ Universal_Receiver.h
@@ -28,10 +28,10 @@ struct RECEIVE_DATA_STRUCTURE
 {
     // put your variable definitions here for the data you want to receive
     // THIS MUST BE EXACTLY THE SAME ON THE OTHER ARDUINO
-    int analogNb;
-    int digitalNb;
-    int analog[NB_MAX_DATA];
-    int digital[NB_MAX_DATA];
+    int16_t analogNb;
+    int16_t digitalNb;
+    int16_t analog[NB_MAX_DATA];
+    int16_t digital[NB_MAX_DATA];
 };
 
 struct SEND_DATA_STRUCTURE
@@ -71,11 +71,11 @@ public:
     int digitalPin[NB_MAX_DATA]; // The Bluetooth state pin is always the first index of the wired sensors.
     bool digitalInputPullup[NB_MAX_DATA];
     bool digitalReversedLogic[NB_MAX_DATA];
-    int analog[NB_MAX_DATA * 2];      // Data received via Bluetooth is stored in the first NB_MAX_DATA indexes, data from wired sensors is stored in the last NB_MAX_DATA indexes.
-    int digital[NB_MAX_DATA * 2];     // Data received via Bluetooth is stored in the first NB_MAX_DATA indexes, data from wired sensors is stored in the last NB_MAX_DATA indexes.
-    int lastAnalog[NB_MAX_DATA * 2];  // Stores the last analog state.
-    int lastDigital[NB_MAX_DATA * 2]; // Stores the last digital state.
-    int threshold[NB_MAX_DATA * 2];   // Holds the threshold for the analog values. If old_data-threshold <= incoming data <= old_data+threshold, then the incoming data is considered the same as the old data.
+    int16_t analog[NB_MAX_DATA * 2];      // Data received via Bluetooth is stored in the first NB_MAX_DATA indexes, data from wired sensors is stored in the last NB_MAX_DATA indexes.
+    int16_t digital[NB_MAX_DATA * 2];     // Data received via Bluetooth is stored in the first NB_MAX_DATA indexes, data from wired sensors is stored in the last NB_MAX_DATA indexes.
+    int16_t lastAnalog[NB_MAX_DATA * 2];  // Stores the last analog state.
+    int16_t lastDigital[NB_MAX_DATA * 2]; // Stores the last digital state.
+    int16_t threshold[NB_MAX_DATA * 2];   // Holds the threshold for the analog values. If old_data-threshold <= incoming data <= old_data+threshold, then the incoming data is considered the same as the old data.
     UPDATED isUpdated;
 
     // create two EasyTransfer objects.
@@ -85,12 +85,12 @@ public:
     RECEIVE_DATA_STRUCTURE rxdata;
     SEND_DATA_STRUCTURE txdata;
 
-    Universal_Receiver(int rx, int tx, int digNb_hw, int anaNb_hw, int digPins[NB_MAX_DATA], int anaPins[NB_MAX_DATA], bool digInputPullup[NB_MAX_DATA], bool digReversedLogic[NB_MAX_DATA], int thresh[NB_MAX_DATA * 2], String btHardware);
-    Universal_Receiver(HardwareSerial *stream, int digNb_hw, int anaNb_hw, int digPins[NB_MAX_DATA], int anaPins[NB_MAX_DATA], bool digInputPullup[NB_MAX_DATA], bool digReversedLogic[NB_MAX_DATA], int thresh[NB_MAX_DATA * 2], String btHardware);
+    Universal_Receiver(int rx, int tx, int digNb_hw, int anaNb_hw, int digPins[NB_MAX_DATA], int anaPins[NB_MAX_DATA], bool digInputPullup[NB_MAX_DATA], bool digReversedLogic[NB_MAX_DATA], int16_t thresh[NB_MAX_DATA * 2], String btHardware);
+    Universal_Receiver(HardwareSerial *stream, int digNb_hw, int anaNb_hw, int digPins[NB_MAX_DATA], int anaPins[NB_MAX_DATA], bool digInputPullup[NB_MAX_DATA], bool digReversedLogic[NB_MAX_DATA], int16_t thresh[NB_MAX_DATA * 2], String btHardware);
     // void setSerialLink(int rx, int tx);
     // void setSerialLink(HardwareSerial *stream);
 
-    void start(int digNb_hw, int anaNb_hw, int digPins[NB_MAX_DATA], int anaPins[NB_MAX_DATA], bool digInputPullup[NB_MAX_DATA], bool digReversedLogic[NB_MAX_DATA], int thresh[NB_MAX_DATA * 2], String btHardware); // Class initializer.
+    void start(int digNb_hw, int anaNb_hw, int digPins[NB_MAX_DATA], int anaPins[NB_MAX_DATA], bool digInputPullup[NB_MAX_DATA], bool digReversedLogic[NB_MAX_DATA], int16_t thresh[NB_MAX_DATA * 2], String btHardware); // Class initializer.
     // Not all pins on the Mega and Mega 2560 support change interrupts, so only the following can be used for RX: 10, 11, 12, 13, 14, 15, 50, 51, 52, 53, A8 (62), A9 (63), A10 (64), A11 (65), A12 (66), A13 (67), A14 (68), A15 (69).
     // Not all pins on the Leonardo and Micro support change interrupts, so only the following can be used for RX: 8, 9, 10, 11, 14 (MISO), 15 (SCK), 16 (MOSI).
     // On Arduino or Genuino 101 RX doesn't work on Pin 13
